requirements.txt
requests
hmac
kafka-python
Flask
python-dotenv
pyproj
pymongo
pyspark
luigi
python-daemon